RAGdemo/requirements.txt
streamlit>=1.30.0
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pypdf>=3.0.0
langchain-text-splitters>=0.0.1
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
python-docx>=1.0.0
